requirements.txt

# Core runtime
streamlit>=1.24.0
weaviate-client>=4.0.0
tiktoken>=0.4.0
pymupdf4llm>=0.2.0
python-docx>=0.8.11
python-dotenv>=1.0.0

# Utilities & testing
requests>=2.31.0
pytest>=7.0.0
